Split view methods out of the Isotope class.
